Make observations page dynamic
