--- src/app/(frontend)/[center]/accidents/error.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-import { ErrorBoundary } from '@/components/ErrorBoundary/ErrorBoundary'
-
-export default ErrorBoundary
